# docs/api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "CheckMyMole API"
  },
  "host" : "prod.api.checkmoleapp.demo-redisys.com",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/users/me/requests" : {
      "get" : {
        "tags" : [ "Requests" ],
        "summary" : "List all of user's requests",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "description" : "How many rows should be skipped",
          "required" : false,
          "type" : "integer",
          "default" : 0
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "How many rows should be returned",
          "required" : false,
          "type" : "integer",
          "default" : 50
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Request"
              }
            }
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "owned.lesions.write" ]
        } ]
      },
      "post" : {
        "tags" : [ "Requests" ],
        "summary" : "Create a new request",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Request to be updated. You can pass a slice of report IDs as the \"reports\" field if you'd like to change their request_id field in bulk. The status field of the reports gets automatically changed to submitted.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Request"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Request"
            }
          },
          "400" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "owned.lesions.write" ]
        } ]
      }
    },
    "/users/me/requests/{id}" : {
      "get" : {
        "tags" : [ "Requests" ],
        "summary" : "Gets a single request with related objects",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the part to update",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "include_reports",
          "in" : "query",
          "description" : "Should reports be included (as \"reports[]\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "temp_urls",
          "in" : "query",
          "description" : "Should reports.photos be signed?",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_lesions",
          "in" : "query",
          "description" : "Should lesions be included (as \"reports[].lesion\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_answers",
          "in" : "query",
          "description" : "Should answers be included (as \"reports[].answers\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "include_questions",
          "in" : "query",
          "description" : "Should questions be included (as \"reports[].answers[].question\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "skip_answer",
          "in" : "query",
          "description" : "Whether answer_text should be skipped",
          "required" : false,
          "type" : "boolean"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Request"
            }
          },
          "404" : {
            "description" : "Request not found"
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "owned.lesions.read" ]
        } ]
      },
      "put" : {
        "tags" : [ "Requests" ],
        "summary" : "Update an existing request",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the part to update",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Request to be updated. You can pass a slice of report IDs as the \"reports\" field if you'd like to change their request_id field in bulk. The status field of the reports gets automatically changed to submitted.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Request"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Request"
            }
          },
          "400" : {
            "description" : "Invalid ID supplied or invalid JSON input passed"
          },
          "404" : {
            "description" : "Request not found"
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "owned.lesions.write" ]
        } ]
      },
      "delete" : {
        "tags" : [ "Requests" ],
        "summary" : "Delete an existing request",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the request to delete",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Request"
            }
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Request not found"
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "owned.lesions.write" ]
        } ]
      }
    },
...
    "/requests" : {
      "get" : {
        "tags" : [ "Requests" ],
        "summary" : "List all of the requests in the system",
        "parameters" : [ {
          "name" : "account_id",
          "in" : "query",
          "description" : "Filter requests by this particular account",
          "required" : false,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "status",
          "in" : "query",
          "description" : "Filter requests with this particular status",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "include_accounts",
          "in" : "query",
          "description" : "Should accounts be included (as the \"account\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "How many rows should be skipped",
          "required" : false,
          "type" : "integer",
          "default" : 0
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "How many rows should be returned",
          "required" : false,
          "type" : "integer",
          "default" : 50
        }, {
          "name" : "skip_answer",
          "in" : "query",
          "description" : "Whether answer_text should be skipped",
          "required" : false,
          "type" : "boolean",
          "default" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Request"
              }
            }
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "requests.read" ]
        } ]
      }
    },
    "/requests/{id}" : {
      "get" : {
        "tags" : [ "Requests" ],
        "summary" : "Get a single request by its ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the request to fetch",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "include_account",
          "in" : "query",
          "description" : "Should the account be included (as the \"account\" field)",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "skip_answer",
          "in" : "query",
          "description" : "Whether answer_text should be skipped",
          "required" : false,
          "type" : "boolean"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Report"
            }
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "requests.read" ]
        } ]
      },
      "put" : {
        "tags" : [ "Requests" ],
        "summary" : "Allows doctors to update a request (including changing answer and state to answered)",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the request to update",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Request to be updated. It can contain a \"notify_msg\" to send a push notification to user's channel.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/RequestWithReports"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Report"
            }
          }
        },
        "security" : [ {
          "checkmoleapp_auth" : [ "requests.respond" ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "checkmoleapp_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "https://auth.checkmoleapp.demo-redisys.com/login",
      "flow" : "implicit",
      "scopes" : {
        "body-parts.write" : "create, update and delete body parts",
        "questions.write" : "create, update and delete questions",
        "owned.lesions.read" : "read lesions, reports and requests owned by the user",
        "owned.lesions.write" : "modify lesions, reports and requests owned by the user",
        "requests.read" : "read all requests in the system",
        "reports.read" : "read all reports in the system",
        "requests.respond" : "update all requests in the system",
        "lesions.read" : "read all lesions in the system"
      }
    }
  },
  "definitions" : {
...
    "Request" : {
      "type" : "object",
      "required" : [ "account_id", "status" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "account_id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "ID of the account that owns the request"
        },
        "status" : {
          "type" : "string",
          "description" : "draft, submitted or answered"
        },
        "answer_text" : {
          "type" : "string",
          "description" : "answer from the doctor"
        },
        "answered_by" : {
          "type" : "string",
          "description" : "name of the doctor"
        },
        "answered_at" : {
          "type" : "integer",
          "description" : "when it was answered"
        },
        "created_at" : {
          "type" : "integer"
        },
        "updated_at" : {
          "type" : "integer"
        }
      },
      "example" : {
        "answer_text" : "answer_text",
        "account_id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "updated_at" : 1,
        "answered_by" : "answered_by",
        "answered_at" : 0,
        "created_at" : 6,
        "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "status" : "status"
      }
    },
    "RequestWithReports" : {
      "type" : "object",
      "required" : [ "status" ],
      "properties" : {
        "status" : {
          "type" : "string",
          "description" : "draft, submitted or answered"
        },
        "answer_text" : {
          "type" : "string",
          "description" : "answer from the doctor"
        },
        "answered_by" : {
          "type" : "string",
          "description" : "name of the doctor"
        },
        "answered_at" : {
          "type" : "integer",
          "description" : "when it was answered"
        },
        "reports" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/RequestWithReports_reports"
          }
        }
      },
      "example" : {
        "answer_text" : "answer_text",
        "reports" : [ {
          "consultation_result" : "consultation_result",
          "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        }, {
          "consultation_result" : "consultation_result",
          "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        } ],
        "answered_by" : "answered_by",
        "answered_at" : 0,
        "status" : "status"
      }
    },
    "Report" : {
      "type" : "object",
      "required" : [ "lesion_id", "photos" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "request_id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "ID of the request that the report belongs to"
        },
        "lesion_id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "ID of the lesion that this report is about"
        },
        "photos" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "description" : "URL of the photo"
          }
        },
        "status" : {
          "type" : "string",
          "description" : "Any value is accepted"
        },
        "consultation_result" : {
          "type" : "string"
        },
        "answers" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Report_answers"
          }
        }
      },
      "example" : {
        "lesion_id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "consultation_result" : "consultation_result",
        "answers" : [ {
          "question" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "answer" : "{}"
        }, {
          "question" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "answer" : "{}"
        } ],
        "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "request_id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "photos" : [ "photos", "photos" ],
        "status" : "status"
      }
    },
    "RequestWithReports_reports" : {
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "consultation_result" : {
          "type" : "string",
          "description" : "Report-specific consultation result"
        }
      },
      "example" : {
        "consultation_result" : "consultation_result",
        "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      }
    },
    "Report_answers" : {
      "properties" : {
        "question" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "ID of the question"
        },
        "answer" : {
          "type" : "object",
          "description" : "JSON value with the answer",
          "properties" : { }
        }
      },
      "example" : {
        "question" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "answer" : "{}"
      }
    }
  }
}